Migrate @now/optipng using TypeScript
